Address coderabbit review

+++ components/core/src/clp_s/indexer/constants.hpp
@@ -2,13 +2,14 @@
 #define CLP_S_INDEXER_CONSTANTS_HPP
 
 #include <array>
+#include <cstddef>
 #include <string_view>
 
 #include "types.hpp"
 
 namespace clp_s::indexer {
 constexpr std::string_view cColumnMetadataTableSuffix = "column_metadata";
-constexpr std::array<std::string_view, static_cast<int>(ColumnMetadataTableFieldIndexes::Length)>
+constexpr std::array<std::string_view, static_cast<size_t>(ColumnMetadataTableFieldIndexes::Length)>
         cColumnMetadataTableFieldNames = {"name", "type"};
 }  // namespace clp_s::indexer
